feat: add indexer start to harness
